Always show the lock icon when an axis is locked
When a user clicks the "lock" icon on an axis to, um, lock the axis, that closed lock icon will now always be shown (so long as the axis remains locked). Previously, the lock icon would only appear when the user moused over the axis area, whether the axis was locked or unlocked. The lock icon will still continue to only be shown on hover when the axis is unlocked.