[WIP] Hacky fix to work around FE-8650 by returning an empty measure name
FE-8650 is a really annoying bug where if you add three VDF sources, enable range chart, set a custom date format, then delete source 1, then the new source 1 (formerly source 2), Immerse crashes. If you delete the sources in any other order, everything is fine.

The source of the crash is that `buildMeasureNameLookup()` tries to access `data.dataBySeries[0].dimensionName`, but `data.dataBySeries[0]` is an empty array, so it throws an "error trying to access dimensionName of undefined" exception and crashes.

I made a super-hacky fix that checks if data.dataBySeries[0] is falsey and, if so, returns an empty string instead of a measure name. This stops Immerse from crashing, but I'm not sure what the full implications of this are. Like, I don't know if I'm causing more bugs here.

mapd3 reviewers, please look at this code and tell me if I'm doing any crimes against correctness here. I *feel* this is the wrong solution, but I can't say for sure one way or another since I'm not super-familiar with this code.